Add a more concise input format for header maps
The `hmaptool` header map builder takes input files with the format

  KEY|VALUE

which is pretty flexible and generic. However, in practice, the
key can always be derived from the value because the value is a
path and the key is the basename with a namespace optionally
prepended.

This change still supports the old format for compatibility with
rules that might be using it, but adds a more compact format
where the input file can be just:

  VALUE

(note the absecne of `|`). The new -`-namespace` flag, is intended
to provide the namespace name for the new format. It has no effect
if using the old format.

The hmap bazel rule has been updated and simplified to make use
of the new format.

**BREAKING CHANGE:** The `flatten_headers` field has been renamed to
`namespace_headers` to make clear what it's doing.